--- cache/nginx/Dockerfile
@@ -1,8 +1,0 @@
-FROM nginx:1.27-alpine
-
-COPY nginx.conf /etc/nginx/nginx.conf
-COPY sites-enabled/ /etc/nginx/sites-enabled/
-
-EXPOSE 8070
-EXPOSE 8080
-EXPOSE 8090
