+++ server/Cargo.toml
@@ -16,7 +16,6 @@ tracing-subscriber = { version = "0.3", features = ["fmt", "env-filter"] }
 axum = "0.7.9"
 tokio = { version="1.41", features=["full"] }
 tower = "0.5.2"
-url = "2.5.4"
 quick_cache = { version = "0.6.16" }
 
 # Kona
@@ -32,6 +31,7 @@ alloy-primitives = { workspace = true, features = ["map", "serde"] }
 alloy-provider = { workspace = true, features = ["reqwest"] }
 alloy-transport = { workspace = true }
 alloy-rpc-client = { workspace = true }
+alloy-transport-http = { workspace = true, features = ["reqwest"] }
 alloy-json-rpc = { workspace = true }
 op-alloy-network = { workspace = true }
